requirements.txt
absl-py==0.9.0
addict==2.4.0
albumentations==1.3.0
backcall==0.2.0
blurhash==1.1.4
boost==0.1
cachetools==4.1.0
certifi==2021.5.30
chardet==3.0.4
charset-normalizer==2.0.12
cmake==3.24.1.1
cycler==0.10.0
dataclasses==0.8
decorator==4.4.2
dill==0.3.4
dlib==19.24.0
eagerpy==0.30.0
efficientnet-pytorch==0.7.1
entrypoints==0.4
foolbox==3.3.3
future==0.18.2
gitdb==4.0.9
GitPython==3.1.18
google-auth==1.15.0
google-auth-oauthlib==0.4.1
grad-cam==1.4.8
graphviz==0.8.4
greenlet==1.1.3
grpcio==1.29.0
idna==2.9
image-quality==1.2.7
imageio==2.8.0
imgaug==0.4.0
importlib-metadata==4.8.3
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
jedi==0.17.2
joblib==1.1.1
jupyter-client==7.1.2
jupyter-core==4.9.2
kiwisolver==1.2.0
libsvm==3.23.0.4
llvmlite==0.36.0
lpips==0.1.4
Markdown==3.2.2
Mastodon.py==1.5.1
matplotlib==3.2.1
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
mmcv-full==1.7.1
multiprocess==0.70.12.2
mxnet==1.6.0
nest-asyncio==1.5.7
networkx==2.4
numba==0.53.1
numpy==1.19.2
oauthlib==3.1.0
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
packaging==21.3
pandas==1.1.5
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
prompt-toolkit==3.0.36
protobuf==3.12.1
ptyprocess==0.7.0
pyaml==21.10.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyDeprecate==0.3.2
Pygments==2.14.0
pynndescent==0.3.3
pyparsing==2.4.7
python-dateutil==2.8.2
python-magic==0.4.27
pytorch-lightning==0.7.1
pytz==2022.4
PyWavelets==1.1.1
PyYAML==6.0
pyzmq==25.1.0
qudida==0.0.4
requests==2.27.1
requests-oauthlib==1.3.0
retinaface-pytorch==0.0.8
rsa==4.0
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.4.1
Shapely==1.8.5.post1
six==1.16.0
smmap==5.0.0
SQLAlchemy==1.4.41
tensorboard==2.2.1
tensorboard-plugin-wit==1.6.0.post3
tensorboardX==1.9
threadpoolctl==3.1.0
tifffile==2020.5.25
timm==0.4.12
torch==1.9.0
torch-fidelity==0.3.0
torchmetrics==0.8.2
torchvision==0.10.0
tornado==6.1
tqdm==4.46.0
traitlets==4.3.3
ttach==0.0.3
typing-extensions==4.1.1
urllib3==1.25.9
wcwidth==0.2.6
Werkzeug==1.0.1
yapf==0.32.0
zipp==3.1.0